--- client/components/filters/filter/utils.js
@@ -1,19 +1,0 @@
-/** @format */
-/**
- * External dependencies
- */
-import { omit } from 'lodash';
-
-export function flatenFilters( filters ) {
-	const allFilters = [];
-	filters.forEach( f => {
-		if ( ! f.subFilters ) {
-			allFilters.push( f );
-		} else {
-			allFilters.push( omit( f, 'subFilters' ) );
-			const subFilters = flatenFilters( f.subFilters );
-			allFilters.push( ...subFilters );
-		}
-	} );
-	return allFilters;
-}
