+++ client/store/reports/utils.js
@@ -9,13 +9,12 @@ import { find, forEach, isNull } from 'lodash';
  * WooCommerce dependencies
  */
 import { appendTimestamp, getCurrentDates, getIntervalForQuery } from '@woocommerce/date';
+import { flattenFilters, getActiveFiltersFromQuery, getUrlKey } from '@woocommerce/navigation';
 
 /**
  * Internal dependencies
  */
 import { MAX_PER_PAGE } from 'store/constants';
-import { getActiveFiltersFromQuery, getUrlKey } from 'components/filters/advanced/utils';
-import { flatenFilters } from 'components/filters/filter/utils';
 import * as couponsConfig from 'analytics/report/coupons/config';
 import * as ordersConfig from 'analytics/report/orders/config';
 import * as productsConfig from 'analytics/report/products/config';
@@ -60,7 +59,7 @@ export function getQueryFromConfig( config, advancedFilters, query ) {
 		);
 	}
 
-	const filter = find( flatenFilters( config.filters ), { value: queryValue } );
+	const filter = find( flattenFilters( config.filters ), { value: queryValue } );
 
 	if ( ! filter ) {
 		return {};
