+++ packages/navigation/README.md
@@ -104,3 +104,65 @@ Updates the query parameters of the current page.
 | query | <code>Object</code> | object of params to be updated. |
 | path | <code>String</code> | Relative path (defaults to current path). |
 | currentQuery | <code>Object</code> | object of current query params (defaults to current querystring). |
+
+## flattenFilters(filters) ⇒ <code>Array</code>
+Collapse an array of filter values with subFilters into a 1-dimensional array.
+
+**Returns**: <code>Array</code> - Flattened array of all filters.
+
+| Param | Type | Description |
+| --- | --- | --- |
+| filters | <code>Array</code> | Set of filters with possible subfilters. |
+
+## getActiveFiltersFromQuery(query, config) ⇒ <code>Array.&lt;activeFilters&gt;</code>
+Given a query object, return an array of activeFilters, if any.
+
+**Returns**: <code>Array.&lt;activeFilters&gt;</code> - - array of activeFilters
+
+| Param | Type | Description |
+| --- | --- | --- |
+| query | <code>object</code> | query oject |
+| config | <code>object</code> | config object |
+
+## getDefaultOptionValue(config, options) ⇒ <code>string</code> \| <code>undefined</code>
+Get the url query key from the filter key and rule.
+
+**Returns**: <code>string</code> \| <code>undefined</code> - - the value of the default option.
+
+| Param | Type | Description |
+| --- | --- | --- |
+| config | <code>object</code> | a filter config object. |
+| options | <code>array</code> | select options. |
+
+## getQueryFromActiveFilters(activeFilters, query, config) ⇒ <code>object</code>
+Given activeFilters, create a new query object to update the url. Use previousFilters to
+Remove unused params.
+
+**Returns**: <code>object</code> - - query object representing the new parameters
+
+| Param | Type | Description |
+| --- | --- | --- |
+| activeFilters | <code>Array.&lt;activeFilters&gt;</code> | activeFilters shown in the UI |
+| query | <code>object</code> | the current url query object |
+| config | <code>object</code> | config object |
+
+## getUrlKey(key, rule) ⇒ <code>string</code>
+Get the url query key from the filter key and rule.
+
+**Returns**: <code>string</code> - - url query key.
+
+| Param | Type | Description |
+| --- | --- | --- |
+| key | <code>string</code> | filter key. |
+| rule | <code>string</code> | filter rule. |
+
+## activeFilter : <code>Object</code>
+Describe activeFilter object.
+
+**Properties**
+
+| Name | Type | Description |
+| --- | --- | --- |
+| key | <code>string</code> | filter key. |
+| [rule] | <code>string</code> | a modifying rule for a filter, eg 'includes' or 'is_not'. |
+| value | <code>string</code> | filter value(s). |
